starting to break up pane service into dedicated files
