+++ services/pane/index.js
@@ -1,24 +1,22 @@
 var _ = require('lodash'),
-  keycode = require('keycode'),
   moment = require('moment'),
   dom = require('@nymag/dom'),
   ds = require('dollar-slice'),
-  edit = require('./edit'),
-  progress = require('./progress'),
-  state = require('./page-state'),
-  site = require('./site'),
-  tpl = require('./tpl'),
-  db = require('../services/edit/db'),
-  datepicker = require('./field-helpers/datepicker'),
-  paneController = require('../controllers/pane'),
-  filterableList = require('./filterable-list'),
-  publishPaneController = require('../controllers/publish-pane'),
-  previewController = require('../controllers/preview-pane'),
-  references = require('./references'),
-  addComponent = require('./components/add-component'),
-  select = require('./components/select'),
-  label = require('./label'),
-  invisibleList = require('./components/invisible-list'),
+  edit = require('../edit'),
+  state = require('../page-state'),
+  site = require('../site'),
+  tpl = require('../tpl'),
+  db = require('../edit/db'),
+  datepicker = require('../field-helpers/datepicker'),
+  paneController = require('../../controllers/pane'),
+  filterableList = require('../filterable-list'),
+  publishPaneController = require('../../controllers/publish-pane'),
+  previewController = require('../../controllers/preview-pane'),
+  references = require('../references'),
+  addComponent = require('../components/add-component'),
+  select = require('../components/select'),
+  label = require('../label'),
+  invisibleList = require('../components/invisible-list'),
   kilnHideClass = 'kiln-hide';
 
 /**
@@ -373,94 +371,6 @@ function openPublish(validation) {
   });
 }
 
-/**
- * create a new page based on the provided ID
- * note: if successful, will redirect to the new page.
- * otherwise, will display an error message
- * @param {string} id
- * @returns {Promise}
- */
-function createPageByType(id) {
-  return edit.createPage(id)
-    .then(function (url) {
-      location.href = url;
-    })
-    .catch(function () {
-      progress.done('error');
-      progress.open('error', 'Error creating new page', true);
-    });
-}
-
-/**
- * add the current page to the list of available pages
- * @param {array} current pages
- * @returns {Function}
- */
-function addCurrentPage(current) {
-  return function (el) {
-    var form = dom.create(`<form class="add-page-form">
-        <input type="text" placeholder="Page Name" />
-        <button type="submit">Add Page To List</button>
-      </form>`),
-      input = dom.find(form, 'input');
-
-    input.addEventListener('keydown', function (e) {
-      var key = keycode(e);
-
-      if (key === 'esc') {
-        dom.replaceElement(form, el);
-      }
-    });
-
-    form.addEventListener('submit', function (e) {
-      var value = input.value,
-        id = _.last(dom.pageUri().split('/'));
-
-      e.preventDefault();
-      progress.start('page');
-      current.push({
-        id: id,
-        title: value
-      });
-
-      return db.save(site.get('prefix') + '/lists/new-pages', current)
-      .then(function () {
-        close();
-        progress.done('page');
-        progress.open('page', `<em>${value}</em> added to new pages list`, true);
-      })
-      .catch(function () {
-        progress.done('error');
-        progress.open('error', 'Error creating new page', true);
-      });
-    });
-
-    dom.replaceElement(el, form);
-    input.focus();
-  };
-}
-
-/**
- * open new page/edit layout dialog pane
- * @returns {Promise}
- */
-function openNewPage() {
-  // /lists/new-pages contains a site-specific array of pages that should be available
-  // to clone, each one having a `id` (the page id) and `title` (the button title) property
-  // note: this shouldn't be cached
-  return db.get(`${site.get('prefix')}/lists/new-pages`)
-    .then(function (items) {
-      var innerEl = filterableList.create(items, {
-        click: createPageByType,
-        add: addCurrentPage(items),
-        addTitle: 'Add Current Page To List'
-      });
-
-      // create pane
-      return open([{header: 'New Page', content: innerEl}]);
-    });
-}
-
 function addPreview(preview) {
   if (preview) {
     return `<span class="error-preview">${preview}</span>`;
@@ -647,10 +557,11 @@ function openComponents() {
 
 module.exports.close = close;
 module.exports.open = open;
-module.exports.openNewPage = openNewPage;
+_.set(window, 'kiln.services.pane', module.exports); // export for plugins
+
+// todo: split these out into separate files
 module.exports.openPublish = openPublish;
 module.exports.openPreview = openPreview;
 module.exports.openAddComponent = openAddComponent;
 module.exports.takeOffEveryZig = takeOffEveryZig;
 module.exports.openComponents = openComponents;
-_.set(window, 'kiln.services.pane', module.exports); // export for plugins
